Add Dependabot configuration file for GitHub Actions and Go modules
This commit adds a new file `.github/dependabot.yml` to the repository. The file specifies the configuration for Dependabot to update dependencies for both GitHub Actions and Go modules. The configuration sets the update schedule to weekly for both package ecosystems. This will ensure that the repository's dependencies stay up to date.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>