feat(gender): Add initialization of random number generator
This commit adds an initialization step for the random number generator in the `generator.go` file. The `init()` function is introduced, which creates a new random number generator using the current time as the seed. This ensures that each run of the program will have a different sequence of random numbers.

The purpose of this change is to improve the randomness of the generated names by providing a more varied set of random numbers for selecting names.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>