test: check if generated name contains family name separator
The test function in the generator_test.go file has been modified to check if the generated name contains a family name separator. If the name does not contain a separator, an error message is printed. Additionally, the code now checks if the generated name is contained in the list of first names and second names. If it is not, an error message is printed.

Signed-off-by: Rodney Osodo <28790446+rodneyosodo@users.noreply.github.com>